+++ packages/orq-rc/RELEASES.md
@@ -5939,3 +5939,13 @@ Based on:
 - [typescript v4.2.0-rc.40] packages/orq-rc
 ### Releases
 - [NPM v4.2.0-rc.40] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.40 - packages/orq-rc
+
+## 2026-01-19 13:16:53
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.688.0 (2.796.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.2.0-rc.41] packages/orq-rc
+### Releases
+- [NPM v4.2.0-rc.41] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.41 - packages/orq-rc